assignment6_monitor_high_alart/requirements.txt
boto3
datetime
os
logging
BotoCoreError
ClientError